requirements.txt
llama-hub==0.0.65
llama-index==0.9.21
chromadb==0.4.10
streamlit==1.29.0
typing==3.7.4.3
python-dotenv==1.0.0
tqdm==4.65.0